requirements.txt
# Basic FastAPI dependencies (EXISTING - keep these)
fastapi==0.104.1
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
uvicorn==0.24.0
psycopg2-binary==2.9.9

# Web scraping dependencies (EXISTING - keep these)
beautifulsoup4==4.12.2
requests==2.31.0

# Basic scientific libraries (EXISTING - keep these)
numpy==1.24.3
scikit-learn==1.3.0

# HuggingFace Hub (EXISTING - keep these)
huggingface-hub==0.17.3

# NEW: OpenAI & Streamlit dependencies (ADD THESE)
openai==1.6.1
streamlit==1.28.0
python-dotenv==1.0.0
plotly==5.17.0

# NEW: Additional utilities for Streamlit
pandas==2.1.0
altair==5.1.0

redis==4.6.0